Move setting presence to ready.js
Makes it more consistent since it happens later in the boot process
