+++ events/shardReady.js
@@ -1,5 +1,3 @@
-import { ActivityType } from "discord.js";
-import globalVars from "../objects/globalVars.json" with { type: "json" };
 import getTime from "../util/getTime.js";
 import stan from "../affairs/stan.js";
 import birthday from "../affairs/birthday.js";
@@ -7,11 +5,6 @@ import birthday from "../affairs/birthday.js";
 export default async (client, id) => {
     try {
         let timestamp = getTime();
-        let presence = initPresence();
-        globalVars.presence = presence;
-        // Set bot status
-        await client.user.setPresence(presence);
-        console.log(`Presence set to "${client.user.presence.activities[0].type} ${client.user.presence.activities[0].name}"`);
         // Start affairs
         stan(client);
         birthday(client);
@@ -23,9 +16,4 @@ export default async (client, id) => {
     } catch (e) {
         console.log(e);
     };
-};
-
-function initPresence() {
-    let presence = { activities: [{ name: 'the lake theme', type: ActivityType.Listening }], status: 'idle' };
-    return presence;
 };
